+++ test/Feature/CleanRegex/Match/MatchPatternTest.php
@@ -239,6 +239,24 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
+    public function shouldForEach_acceptKey()
+    {
+        // given
+        $arguments = [];
+
+        // when
+        pattern('\w+')->match('Foo, Bar, Cat')
+            ->forEach(function (string $argument, int $index) use (&$arguments) {
+                $arguments[$argument] = $index;
+            });
+
+        // then
+        $this->assertSame(['Foo' => 0, 'Bar' => 1, 'Cat' => 2], $arguments);
+    }
+
+    /**
+     * @test
+     */
     public function shouldGet_offsets()
     {
         // given